python-projects/01-text-generator/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
anthropic>=0.18.0
openai>=1.12.0
